Manually deploy PriceFeed
We're running into a problem where calling `deployments.fixture` is breaking other tets due to a known bug in hardhat.

In the short term, we fix this by manually deploying it. Long-term, we can create a universal test deployment and use that for everything, pulling in acre's helper functions for snapshots