Provide an echidna surface to efficiently deposit to the stability pool
Echidna development can be tricky; it generates random parameters, so if one of
those params is a uint or uint256, and that represents how much money you want
to move, the number it generates is going to be too high the vast majority of
the time. If this is the only way to hit a particular code path, you have to get very lucky to hit it.

Better is to provide a "safe" function that transforms the random data into
reasonable inputs. For instance, if you're generating a uint256 but are trying
to transfer an amount less than your balance, you can do `input % balance` to
have a bounded random number.